backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4


sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
geopy==2.4.1
folium==0.15.0
networkx
geojson
geopy==2.4.1
aiohttp==3.9.1
Pillow
scipy 
numpy
openai==1.54.3
azure-search-documents==11.4.0
azure-core==1.31.0 
pydantic[email]==2.5.0
pandas==2.3.0
httpx==0.27.0
azure-cognitiveservices-speech==1.44.0
azure-cognitiveservices-vision-customvision==3.1.1
# 기존 requirements에 추가
psycopg2-binary==2.9.9

pydub